+++ 1_20_core/src/main/java/com/bteconosur/core/util/PluginRegistry.java
@@ -1,4 +1,4 @@
-package com.bteconosur.core.utils;
+package com.bteconosur.core.util;
 
 import com.bteconosur.core.BteConoSurCore;
 import org.bukkit.Bukkit;
